+++ lib/representable/decorator/coercion.rb
@@ -1,4 +1,0 @@
-require 'representable/coercion'
-
-# TODO: deprecate for 1.7.
-Representable::Decorator::Coercion = Representable::Coercion